+++ base/Cat.ts
@@ -1,0 +1,9 @@
+import { Animal } from "./animal";
+class Cat {
+    name: string
+    wool: boolean
+    constructor(name: string, wool: boolean){
+        this.name = name
+        this.wool = true
+    }
+}
