+++ base/Tiger.ts
@@ -1,0 +1,9 @@
+import { Animal } from "./animal";
+class Tiger {
+   name: string
+   teeth: boolean
+   constructor(name: string, teerh: boolean){
+       this.name = name
+       this.teeth = true
+   }
+}
