+++ base/package.json
@@ -3,7 +3,7 @@
   "version": "1.0.0",
   "main": "index.js",
   "scripts": {
-    "test": "ts-node index.ts"
+    "test": "ts-node index.ts" 
   },
   "author": "",
   "license": "ISC",
